revert last patch because Py_ssize_t is undefined in some (older?)
2008-06-28  Michael Natterer  <mitch@gimp.org>

	* pygimp-pdb.c:
	* pygimp-tile.c:
	* pygimp-colors.c: revert last patch because Py_ssize_t is
	undefined in some (older?) pythons. Needs more investigation.


svn path=/trunk/; revision=26016

+++ plug-ins/pygimp/pygimp-pdb.c
@@ -820,7 +820,7 @@ pf_call(PyGimpPDBFunction *self, PyObject *args, PyObject *kwargs)
 #endif
 
     if (kwargs) {
-        Py_ssize_t len, pos;
+        int len, pos;
         PyObject *key, *val;
 
         len = PyDict_Size(kwargs);
